// apps/web-services/src/models/content.model.ts
export interface IContent {
  appIcon: string
  artifactUrl: string
  averageRating?: number | string | null
  body?: string
  certificationList: IContent[]
  certificationStatus?: TCertificationStatus
  certificationSubmissionDate?: string
  certificationUrl: string
  children: IContent[]
  collections?: IContent[]
  complexityLevel: string
  contentType: TContentType
  contentUrlAtSource: string
  creatorContacts: ICreator[]
  creatorDetails: ICreator[]
  description: string
  displayContentType?: string // For UI
  downloadUrl?: string
  duration: number
  exclusiveContent?: boolean
  hasAccess: boolean
  idealScreenSize?: string
  identifier: string
  introductoryVideo?: string
  introductoryVideoIcon?: string
  isExternal: string | boolean // made as optional boolean for UI
  isIframeSupported: 'Yes' | 'No' | 'Maybe'
  lastUpdatedOn: string
  learningMode?: TLearningMode
  learningObjective: string
  locale: string
  me_totalSessionsCount: number
  mimeType: EMimeTypes
  minLexVersion?: string
  msArtifactDetails?: IMSArtifactDetails
  name: string
  nextCertificationAttemptDate?: string
  playgroundInstructions?: string
  playgroundResources?: IResourcePlayground[]
  postContents?: IPrePostContent[]
  preContents?: IPrePostContent[]
  preRequisites: string
  proctorUrl?: string
  progress?: number
  recentCertificationAttemptScore?: number
  recommendationReasons?: string[]
  registrationInstructions?: string
  resourceCategory?: string[]
  resourceType: string
  skills: ISkill[]
  softwareRequirements?: IResourceDetail[]
  sourceName: string
  ssoEnabled?: boolean
  subtitle?: string
  subTitles?: ISubtitle[]
  status?: string
  systemRequirements?: string[]
  tags: ITag[]
  topics: IContentTopic[]
  totalRating?: number
  track: ITrack[]
}

export type TLearningMode = 'Self-Paced' | 'Instructor-Led'
export type TContentType =
  | 'Learning Path'
  | 'Course'
  | 'Collection'
  | 'Channel'
  | 'Resource'
  | 'Knowledge Board'
  | 'Knowledge Artifact'
  | 'Learning Journeys'
  | 'Playlist'
  | 'Goal'

export enum ECollectionTypes {
  PLAYLIST = 'playlist',
  GOAL = 'goal',
}

type TCertificationStatus = 'ongoing' | 'passed' | 'canAttempt' | 'cannotAttempt'

export enum EMimeTypes {
  COLLECTION = 'application/vnd.ekstep.content-collection',
  HTML = 'application/html',
  ILP_FP = 'application/ilpfp',
  IAP = 'application/iap-assessment',
  MP3 = 'audio/mpeg',
  MP4 = 'video/mp4',
  M3U8 = 'application/x-mpegURL',
  INTERACTIVE_VIDEO = 'video/interactive',
  PDF = 'application/pdf',
  QUIZ = 'application/quiz',
  DRAG_DROP = 'application/drag-drop',
  HTML_PICKER = 'application/htmlpicker',
  WEB_MODULE = 'application/web-module',
  YOUTUBE = 'video/x-youtube',
  HANDS_ON = 'application/integrated-hands-on',
  CLASS_DIAGRAM = 'application/class-diagram',
  COLLECTION_RESOURCE = 'resource/collection',
  // Added on UI Only
  CERTIFICATION = 'application/certification',
  PLAYLIST = 'application/playlist',

  UNKNOWN = 'application/unknown',
}

interface ITag {
  id: string
  type: string
  value: string
}
interface IMSArtifactDetails {
  channelId: string
  videoId: string
}
interface IClient {
  displayName: string
  id: string
  name: string
}
interface ISubtitle {
  srclang: string
  label: string
  url: string
}
interface IPrePostContent {
  identifier: string
  name: string
}
interface IResourceDetail {
  title?: string
  url?: string
}
interface IResourcePlayground {
  appIcon: string
  artifactUrl: string
  identifier: string
  name: string
}
interface ITrack {
  id: string
  name: string
  status: string
  visibility: string
}
interface ISkill {
  id: string
  category: string
  skill: string
  name: string
}
export interface ICreator {
  id: string
  name: string
  email: string
}
export interface IContentTopic {
  identifier: string
  name: string
}
// ...
export interface IContentMinimal {
  appIcon: string
  artifactUrl: string
  complexityLevel: string
  contentType: TContentType
  description: string
  creatorDetails: ICreator[]
  displayContentType: string
  duration: number
  identifier: string
  learningMode?: TLearningMode
  mimeType: EMimeTypes
  name: string
  resourceType?: string
  status?: string
}